Hack around some apparent brokenness in some versions of IO::File.
To do file IO on a scalar variable, we need to pass a ref to open(), but the
"filename" from IO::File is only passed unmolested if a : is detected in the
mode parameter.  So, we force a : on the end of the mode.  Otherwise, we end up
with broken tests, and a bunch of files named SCALAR(0x12345678)