make_singlepart() now deletes all Content-* attributes when flattening to single-part.
This is an odd edge-case.  If you flatten a zero-part multipart message into a
single part (admittedly, a silly thing to do, but it's documented as a feature
of MIME-tools...), we need to delete all content-type, content-encoding, etc,
attributes, as we really don't want the encodings and MIME boundaries and such
to carry over to the new empty text/plain part.

Fixes rt.cpan.org bug 37139