Cleanup: Remove IO::ScalarArray from MIME-tools.
Reading to a filehandle opened on a scalar, and then splitting on
line-beginnings is actually 100% faster than IO::ScalarArray and has none of
its evil baggage, though it can be less memory efficient.  So, we use this
method for preamble and epilogue via read_lines(), but remove read_lines()
entirely in the critical path.

Passing an array in to parse_data() is now deprecated.  For backwards
compatibility, we create a temporary scalar with all lines of the array joined,
and parse that instead.