Fix bizzare edgecase in preamble handling.
This problem was created by the removal of IO-Stringy.  The problem is, when
we parse:

SomeHeader: foo

--mimeboundary

We are expecting a preamble of '', and the parser finds us one.  But, when the
parser hands that off to the caller (by printing '' to a filehandle opened on
a scalar) that print generates nothing, whereas printing same to an
IO::ScalarArray object creates an empty line.

There's probably a better way, but it would involve rewriting much of the
line-reading code.